+++ docs/guides/changelog.md
@@ -11,6 +11,7 @@ description: |-
 
 - Add `sdwan_application_aware_routing_policy_definition` resource and data source
 - Add `sdwan_cflowd_policy_definition` resource and data source
+- Add `sdwan_traffic_data_policy_definition` resource and data source
 
 ## 0.2.1